GFX: draw_rectangle into draw/fill_rect + 'Color' type

+++ src/kernel/init.c
@@ -95,7 +95,7 @@ void kernel_setup(multiboot_info_t *header, unsigned int magic)
     vfs_setroot(vfsinterface);
 
     kprintf("Starting Compositor Server");
-    process_create("Compositor Server", __compositor_main, paging_kernel_pgdir());
+    process_create("Compositor Server", (uint32_t) __compositor_main, paging_kernel_pgdir());
 
     kprintf("All done. Ready to start!\n");
 }
